requirements.txt
Flask==2.1.2
Flask_MySQLdb==1.0.1
mysql_connector_repackaged==0.3.1
PyYAML==6.0

